requirements.txt
hikari[speedups]>=2.0.0.dev114
hikari-lightbulb>=2.3.1
hikari-miru>=2.0.1
python-dotenv>=0.21.0